rag-microservice/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-microservice"
version = "0.1.0"
description = "Retrieval-Augmented Generation (RAG) microservice with vector search and LLM response generation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.0",
    "uvicorn[standard]>=0.35.0",
    "pydantic>=2.11.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.5.0",
    "chromadb>=1.0.0",
    "ollama>=0.5.0",
    "prometheus-fastapi-instrumentator>=6.1.0",
    "httpx>=0.28.0",
]

[tool.setuptools.packages.find]
include = ["app*", "core*"]

[dependency-groups]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=1.3.0",
    "httpx>=0.28.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
